daemon: split find support to its own api_*.go files
