+++ daemon/api.go
@@ -28,9 +28,7 @@ import (
 	"io/ioutil"
 	"mime"
 	"mime/multipart"
-	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -42,7 +40,6 @@ import (
 	"github.com/snapcore/snapd/asserts/snapasserts"
 	"github.com/snapcore/snapd/client"
 	"github.com/snapcore/snapd/dirs"
-	"github.com/snapcore/snapd/httputil"
 	"github.com/snapcore/snapd/i18n"
 	"github.com/snapcore/snapd/logger"
 	"github.com/snapcore/snapd/osutil"
@@ -108,12 +105,6 @@ var (
 		Path:   "/v2/icons/{name}/icon",
 		UserOK: true,
 		GET:    appIconGet,
-	}
-
-	findCmd = &Command{
-		Path:   "/v2/find",
-		UserOK: true,
-		GET:    searchStore,
 	}
 
 	snapsCmd = &Command{
@@ -257,213 +248,6 @@ func getSections(c *Command, r *http.Request, user *auth.UserState) Response {
 	return SyncResponse(sections, nil)
 }
 
-func searchStore(c *Command, r *http.Request, user *auth.UserState) Response {
-	route := c.d.router.Get(snapCmd.Path)
-	if route == nil {
-		return InternalError("cannot find route for snaps")
-	}
-	query := r.URL.Query()
-	q := query.Get("q")
-	commonID := query.Get("common-id")
-	// TODO: support both "category" (search v2) and "section"
-	section := query.Get("section")
-	name := query.Get("name")
-	scope := query.Get("scope")
-	private := false
-	prefix := false
-
-	if sel := query.Get("select"); sel != "" {
-		switch sel {
-		case "refresh":
-			if commonID != "" {
-				return BadRequest("cannot use 'common-id' with 'select=refresh'")
-			}
-			if name != "" {
-				return BadRequest("cannot use 'name' with 'select=refresh'")
-			}
-			if q != "" {
-				return BadRequest("cannot use 'q' with 'select=refresh'")
-			}
-			return storeUpdates(c, r, user)
-		case "private":
-			private = true
-		}
-	}
-
-	if name != "" {
-		if q != "" {
-			return BadRequest("cannot use 'q' and 'name' together")
-		}
-		if commonID != "" {
-			return BadRequest("cannot use 'common-id' and 'name' together")
-		}
-
-		if name[len(name)-1] != '*' {
-			return findOne(c, r, user, name)
-		}
-
-		prefix = true
-		q = name[:len(name)-1]
-	}
-
-	if commonID != "" && q != "" {
-		return BadRequest("cannot use 'common-id' and 'q' together")
-	}
-
-	theStore := getStore(c)
-	ctx := store.WithClientUserAgent(r.Context(), r)
-	found, err := theStore.Find(ctx, &store.Search{
-		Query:    q,
-		Prefix:   prefix,
-		CommonID: commonID,
-		Category: section,
-		Private:  private,
-		Scope:    scope,
-	}, user)
-	switch err {
-	case nil:
-		// pass
-	case store.ErrBadQuery:
-		return SyncResponse(&resp{
-			Type:   ResponseTypeError,
-			Result: &errorResult{Message: err.Error(), Kind: client.ErrorKindBadQuery},
-			Status: 400,
-		}, nil)
-	case store.ErrUnauthenticated, store.ErrInvalidCredentials:
-		return Unauthorized(err.Error())
-	default:
-		if e, ok := err.(*url.Error); ok {
-			if neterr, ok := e.Err.(*net.OpError); ok {
-				if dnserr, ok := neterr.Err.(*net.DNSError); ok {
-					return SyncResponse(&resp{
-						Type:   ResponseTypeError,
-						Result: &errorResult{Message: dnserr.Error(), Kind: client.ErrorKindDNSFailure},
-						Status: 400,
-					}, nil)
-				}
-			}
-		}
-		if e, ok := err.(net.Error); ok && e.Timeout() {
-			return SyncResponse(&resp{
-				Type:   ResponseTypeError,
-				Result: &errorResult{Message: err.Error(), Kind: client.ErrorKindNetworkTimeout},
-				Status: 400,
-			}, nil)
-		}
-		if e, ok := err.(*httputil.PersistentNetworkError); ok {
-			return SyncResponse(&resp{
-				Type:   ResponseTypeError,
-				Result: &errorResult{Message: e.Error(), Kind: client.ErrorKindDNSFailure},
-				Status: 400,
-			}, nil)
-		}
-
-		return InternalError("%v", err)
-	}
-
-	meta := &Meta{
-		SuggestedCurrency: theStore.SuggestedCurrency(),
-		Sources:           []string{"store"},
-	}
-
-	return sendStorePackages(route, meta, found)
-}
-
-func findOne(c *Command, r *http.Request, user *auth.UserState, name string) Response {
-	if err := snap.ValidateName(name); err != nil {
-		return BadRequest(err.Error())
-	}
-
-	theStore := getStore(c)
-	spec := store.SnapSpec{
-		Name: name,
-	}
-	ctx := store.WithClientUserAgent(r.Context(), r)
-	snapInfo, err := theStore.SnapInfo(ctx, spec, user)
-	switch err {
-	case nil:
-		// pass
-	case store.ErrInvalidCredentials:
-		return Unauthorized("%v", err)
-	case store.ErrSnapNotFound:
-		return SnapNotFound(name, err)
-	default:
-		return InternalError("%v", err)
-	}
-
-	meta := &Meta{
-		SuggestedCurrency: theStore.SuggestedCurrency(),
-		Sources:           []string{"store"},
-	}
-
-	results := make([]*json.RawMessage, 1)
-	data, err := json.Marshal(webify(mapRemote(snapInfo), r.URL.String()))
-	if err != nil {
-		return InternalError(err.Error())
-	}
-	results[0] = (*json.RawMessage)(&data)
-	return SyncResponse(results, meta)
-}
-
-func shouldSearchStore(r *http.Request) bool {
-	// we should jump to the old behaviour iff q is given, or if
-	// sources is given and either empty or contains the word
-	// 'store'.  Otherwise, local results only.
-
-	query := r.URL.Query()
-
-	if _, ok := query["q"]; ok {
-		logger.Debugf("use of obsolete \"q\" parameter: %q", r.URL)
-		return true
-	}
-
-	if src, ok := query["sources"]; ok {
-		logger.Debugf("use of obsolete \"sources\" parameter: %q", r.URL)
-		if len(src) == 0 || strings.Contains(src[0], "store") {
-			return true
-		}
-	}
-
-	return false
-}
-
-func storeUpdates(c *Command, r *http.Request, user *auth.UserState) Response {
-	route := c.d.router.Get(snapCmd.Path)
-	if route == nil {
-		return InternalError("cannot find route for snaps")
-	}
-
-	state := c.d.overlord.State()
-	state.Lock()
-	updates, err := snapstateRefreshCandidates(state, user)
-	state.Unlock()
-	if err != nil {
-		return InternalError("cannot list updates: %v", err)
-	}
-
-	return sendStorePackages(route, nil, updates)
-}
-
-func sendStorePackages(route *mux.Route, meta *Meta, found []*snap.Info) Response {
-	results := make([]*json.RawMessage, 0, len(found))
-	for _, x := range found {
-		url, err := route.URL("name", x.InstanceName())
-		if err != nil {
-			logger.Noticef("Cannot build URL for snap %q revision %s: %v", x.InstanceName(), x.Revision, err)
-			continue
-		}
-
-		data, err := json.Marshal(webify(mapRemote(x), url.String()))
-		if err != nil {
-			return InternalError("%v", err)
-		}
-		raw := json.RawMessage(data)
-		results = append(results, &raw)
-	}
-
-	return SyncResponse(results, meta)
-}
-
 // plural!
 func getSnapsInfo(c *Command, r *http.Request, user *auth.UserState) Response {
 
@@ -524,6 +308,28 @@ func getSnapsInfo(c *Command, r *http.Request, user *auth.UserState) Response {
 	}
 
 	return SyncResponse(results, &Meta{Sources: []string{"local"}})
+}
+
+func shouldSearchStore(r *http.Request) bool {
+	// we should jump to the old behaviour iff q is given, or if
+	// sources is given and either empty or contains the word
+	// 'store'.  Otherwise, local results only.
+
+	query := r.URL.Query()
+
+	if _, ok := query["q"]; ok {
+		logger.Debugf("use of obsolete \"q\" parameter: %q", r.URL)
+		return true
+	}
+
+	if src, ok := query["sources"]; ok {
+		logger.Debugf("use of obsolete \"sources\" parameter: %q", r.URL)
+		if len(src) == 0 || strings.Contains(src[0], "store") {
+			return true
+		}
+	}
+
+	return false
 }
 
 // licenseData holds details about the snap license, and may be
