+++ daemon/api_test.go
@@ -308,26 +308,6 @@ UnitFileState=enabled
 	c.Check(rsp.Result, check.DeepEquals, expected.Result)
 }
 
-func (s *apiSuite) TestSnapInfoWithAuth(c *check.C) {
-	s.daemon(c)
-
-	state := snapCmd.d.overlord.State()
-	state.Lock()
-	user, err := auth.NewUser(state, "username", "email@test.com", "macaroon", []string{"discharge"})
-	state.Unlock()
-	c.Check(err, check.IsNil)
-
-	req, err := http.NewRequest("GET", "/v2/find/?q=name:gfoo", nil)
-	c.Assert(err, check.IsNil)
-
-	c.Assert(s.user, check.IsNil)
-
-	_, ok := searchStore(findCmd, req, user).(*resp)
-	c.Assert(ok, check.Equals, true)
-	// ensure user was set
-	c.Assert(s.user, check.DeepEquals, user)
-}
-
 func (s *apiSuite) TestSnapInfoNotFound(c *check.C) {
 	req, err := http.NewRequest("GET", "/v2/snaps/gfoo", nil)
 	c.Assert(err, check.IsNil)
@@ -890,6 +870,20 @@ func (s *apiSuite) TestSnapsInfoOnePerIntegrationAllSome(c *check.C) {
 	s.checkSnapInfoOnePerIntegration(c, true, []string{"foo", "baz"})
 }
 
+func snapList(rawSnaps interface{}) []map[string]interface{} {
+	snaps := make([]map[string]interface{}, len(rawSnaps.([]*json.RawMessage)))
+	for i, raw := range rawSnaps.([]*json.RawMessage) {
+		err := json.Unmarshal([]byte(*raw), &snaps[i])
+		if err != nil {
+			panic(err)
+		}
+	}
+	return snaps
+}
+
+// XXX this should go away
+var SnapList = snapList
+
 func (s *apiSuite) checkSnapInfoOnePerIntegration(c *check.C, all bool, names []string) {
 	d := s.daemon(c)
 
@@ -1080,467 +1074,6 @@ func (s *apiSuite) TestSnapsInfoAll(c *check.C) {
 	}
 }
 
-func (s *apiSuite) TestFind(c *check.C) {
-	s.daemon(c)
-
-	s.suggestedCurrency = "EUR"
-
-	s.rsnaps = []*snap.Info{{
-		SideInfo: snap.SideInfo{
-			RealName: "store",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-
-	req, err := http.NewRequest("GET", "/v2/find?q=hi", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := searchStore(findCmd, req, nil).(*resp)
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Assert(snaps[0]["name"], check.Equals, "store")
-	c.Check(snaps[0]["prices"], check.IsNil)
-	c.Check(snaps[0]["channels"], check.IsNil)
-
-	c.Check(rsp.SuggestedCurrency, check.Equals, "EUR")
-
-	c.Check(s.storeSearch, check.DeepEquals, store.Search{Query: "hi"})
-	c.Check(s.currentSnaps, check.HasLen, 0)
-	c.Check(s.actions, check.HasLen, 0)
-}
-
-func (s *apiSuite) TestFindRefreshes(c *check.C) {
-	snapstateRefreshCandidates = snapstate.RefreshCandidates
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		SideInfo: snap.SideInfo{
-			RealName: "store",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-	s.MockSnap(c, "name: store\nversion: 1.0")
-
-	req, err := http.NewRequest("GET", "/v2/find?select=refresh", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := searchStore(findCmd, req, nil).(*resp)
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Assert(snaps[0]["name"], check.Equals, "store")
-	c.Check(s.currentSnaps, check.HasLen, 1)
-	c.Check(s.actions, check.HasLen, 1)
-}
-
-func (s *apiSuite) TestFindRefreshSideloaded(c *check.C) {
-	snapstateRefreshCandidates = snapstate.RefreshCandidates
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		SideInfo: snap.SideInfo{
-			RealName: "store",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-
-	s.MockSnap(c, "name: store\nversion: 1.0")
-
-	var snapst snapstate.SnapState
-	st := s.d.overlord.State()
-	st.Lock()
-	err := snapstate.Get(st, "store", &snapst)
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-	c.Assert(snapst.Sequence, check.HasLen, 1)
-
-	// clear the snapid
-	snapst.Sequence[0].SnapID = ""
-	st.Lock()
-	snapstate.Set(st, "store", &snapst)
-	st.Unlock()
-
-	req, err := http.NewRequest("GET", "/v2/find?select=refresh", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := searchStore(findCmd, req, nil).(*resp)
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 0)
-	c.Check(s.currentSnaps, check.HasLen, 0)
-	c.Check(s.actions, check.HasLen, 0)
-}
-
-func (s *apiSuite) TestFindPrivate(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{}
-
-	req, err := http.NewRequest("GET", "/v2/find?q=foo&select=private", nil)
-	c.Assert(err, check.IsNil)
-
-	_ = searchStore(findCmd, req, nil).(*resp)
-
-	c.Check(s.storeSearch, check.DeepEquals, store.Search{
-		Query:   "foo",
-		Private: true,
-	})
-}
-
-func (s *apiSuite) TestFindUserAgentContextCreated(c *check.C) {
-	s.daemon(c)
-
-	req, err := http.NewRequest("GET", "/v2/find", nil)
-	c.Assert(err, check.IsNil)
-	req.Header.Add("User-Agent", "some-agent/1.0")
-
-	_ = searchStore(findCmd, req, nil).(*resp)
-
-	c.Check(store.ClientUserAgent(s.ctx), check.Equals, "some-agent/1.0")
-}
-
-func (s *apiSuite) TestFindOneUserAgentContextCreated(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		SnapType: snap.TypeApp,
-		Version:  "v2",
-		SideInfo: snap.SideInfo{
-			RealName: "banana",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-	req, err := http.NewRequest("GET", "/v2/find?name=foo", nil)
-	c.Assert(err, check.IsNil)
-	req.Header.Add("User-Agent", "some-agent/1.0")
-
-	_ = searchStore(findCmd, req, nil).(*resp)
-
-	c.Check(store.ClientUserAgent(s.ctx), check.Equals, "some-agent/1.0")
-}
-
-func (s *apiSuite) TestFindPrefix(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{}
-
-	req, err := http.NewRequest("GET", "/v2/find?name=foo*", nil)
-	c.Assert(err, check.IsNil)
-
-	_ = searchStore(findCmd, req, nil).(*resp)
-
-	c.Check(s.storeSearch, check.DeepEquals, store.Search{Query: "foo", Prefix: true})
-}
-
-func (s *apiSuite) TestFindSection(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{}
-
-	req, err := http.NewRequest("GET", "/v2/find?q=foo&section=bar", nil)
-	c.Assert(err, check.IsNil)
-
-	_ = searchStore(findCmd, req, nil).(*resp)
-
-	c.Check(s.storeSearch, check.DeepEquals, store.Search{
-		Query:    "foo",
-		Category: "bar",
-	})
-}
-
-func (s *apiSuite) TestFindScope(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{}
-
-	req, err := http.NewRequest("GET", "/v2/find?q=foo&scope=creep", nil)
-	c.Assert(err, check.IsNil)
-
-	_ = searchStore(findCmd, req, nil).(*resp)
-
-	c.Check(s.storeSearch, check.DeepEquals, store.Search{
-		Query: "foo",
-		Scope: "creep",
-	})
-}
-
-func (s *apiSuite) TestFindCommonID(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		SideInfo: snap.SideInfo{
-			RealName: "store",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-		CommonIDs: []string{"org.foo"},
-	}}
-	s.MockSnap(c, "name: store\nversion: 1.0")
-
-	req, err := http.NewRequest("GET", "/v2/find?name=foo", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := searchStore(findCmd, req, nil).(*resp)
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Check(snaps[0]["common-ids"], check.DeepEquals, []interface{}{"org.foo"})
-}
-
-func (s *apiSuite) TestFindByCommonID(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		SideInfo: snap.SideInfo{
-			RealName: "store",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-		CommonIDs: []string{"org.foo"},
-	}}
-	s.MockSnap(c, "name: store\nversion: 1.0")
-
-	req, err := http.NewRequest("GET", "/v2/find?common-id=org.foo", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := searchStore(findCmd, req, nil).(*resp)
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Check(s.storeSearch, check.DeepEquals, store.Search{CommonID: "org.foo"})
-}
-
-func (s *apiSuite) TestFindOne(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		SideInfo: snap.SideInfo{
-			RealName: "store",
-		},
-		Base: "base0",
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "verified",
-		},
-		Channels: map[string]*snap.ChannelSnapInfo{
-			"stable": {
-				Revision: snap.R(42),
-			},
-		},
-	}}
-	s.MockSnap(c, "name: store\nversion: 1.0")
-
-	req, err := http.NewRequest("GET", "/v2/find?name=foo", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := searchStore(findCmd, req, nil).(*resp)
-
-	c.Check(s.storeSearch, check.DeepEquals, store.Search{})
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Check(snaps[0]["name"], check.Equals, "store")
-	c.Check(snaps[0]["base"], check.Equals, "base0")
-	c.Check(snaps[0]["publisher"], check.DeepEquals, map[string]interface{}{
-		"id":           "foo-id",
-		"username":     "foo",
-		"display-name": "Foo",
-		"validation":   "verified",
-	})
-	m := snaps[0]["channels"].(map[string]interface{})["stable"].(map[string]interface{})
-
-	c.Check(m["revision"], check.Equals, "42")
-}
-
-func (s *apiSuite) TestFindOneNotFound(c *check.C) {
-	s.daemon(c)
-
-	s.err = store.ErrSnapNotFound
-	s.MockSnap(c, "name: store\nversion: 1.0")
-
-	req, err := http.NewRequest("GET", "/v2/find?name=foo", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := searchStore(findCmd, req, nil).(*resp)
-
-	c.Check(s.storeSearch, check.DeepEquals, store.Search{})
-	c.Check(rsp.Status, check.Equals, 404)
-}
-
-func (s *apiSuite) TestFindRefreshNotOther(c *check.C) {
-	for _, other := range []string{"name", "q", "common-id"} {
-		req, err := http.NewRequest("GET", "/v2/find?select=refresh&"+other+"=foo*", nil)
-		c.Assert(err, check.IsNil)
-
-		rsp := searchStore(findCmd, req, nil).(*resp)
-		c.Check(rsp.Type, check.Equals, ResponseTypeError)
-		c.Check(rsp.Status, check.Equals, 400)
-		c.Check(rsp.Result.(*errorResult).Message, check.Equals, "cannot use '"+other+"' with 'select=refresh'")
-	}
-}
-
-func (s *apiSuite) TestFindNotTogether(c *check.C) {
-	queries := map[string]string{"q": "foo", "name": "foo*", "common-id": "foo"}
-	for ki, vi := range queries {
-		for kj, vj := range queries {
-			if ki == kj {
-				continue
-			}
-
-			req, err := http.NewRequest("GET", fmt.Sprintf("/v2/find?%s=%s&%s=%s", ki, vi, kj, vj), nil)
-			c.Assert(err, check.IsNil)
-
-			rsp := searchStore(findCmd, req, nil).(*resp)
-			c.Check(rsp.Type, check.Equals, ResponseTypeError)
-			c.Check(rsp.Status, check.Equals, 400)
-			exp1 := "cannot use '" + ki + "' and '" + kj + "' together"
-			exp2 := "cannot use '" + kj + "' and '" + ki + "' together"
-			c.Check(rsp.Result.(*errorResult).Message, check.Matches, exp1+"|"+exp2)
-		}
-	}
-}
-
-func (s *apiSuite) TestFindBadQueryReturnsCorrectErrorKind(c *check.C) {
-	s.daemon(c)
-
-	s.err = store.ErrBadQuery
-	req, err := http.NewRequest("GET", "/v2/find?q=return-bad-query-please", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := searchStore(findCmd, req, nil).(*resp)
-	c.Check(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Status, check.Equals, 400)
-	c.Check(rsp.Result.(*errorResult).Message, check.Matches, "bad query")
-	c.Check(rsp.Result.(*errorResult).Kind, check.Equals, client.ErrorKindBadQuery)
-}
-
-func (s *apiSuite) TestFindPriced(c *check.C) {
-	s.daemon(c)
-
-	s.suggestedCurrency = "GBP"
-
-	s.rsnaps = []*snap.Info{{
-		SnapType: snap.TypeApp,
-		Version:  "v2",
-		Prices: map[string]float64{
-			"GBP": 1.23,
-			"EUR": 2.34,
-		},
-		MustBuy: true,
-		SideInfo: snap.SideInfo{
-			RealName: "banana",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-
-	req, err := http.NewRequest("GET", "/v2/find?q=banana&channel=stable", nil)
-	c.Assert(err, check.IsNil)
-	rsp, ok := searchStore(findCmd, req, nil).(*resp)
-	c.Assert(ok, check.Equals, true)
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-
-	snap := snaps[0]
-	c.Check(snap["name"], check.Equals, "banana")
-	c.Check(snap["prices"], check.DeepEquals, map[string]interface{}{
-		"EUR": 2.34,
-		"GBP": 1.23,
-	})
-	c.Check(snap["status"], check.Equals, "priced")
-
-	c.Check(rsp.SuggestedCurrency, check.Equals, "GBP")
-}
-
-func (s *apiSuite) TestFindScreenshotted(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		SnapType: snap.TypeApp,
-		Version:  "v2",
-		Media: []snap.MediaInfo{
-			{
-				Type:   "screenshot",
-				URL:    "http://example.com/screenshot.png",
-				Width:  800,
-				Height: 1280,
-			},
-			{
-				Type: "screenshot",
-				URL:  "http://example.com/screenshot2.png",
-			},
-		},
-		MustBuy: true,
-		SideInfo: snap.SideInfo{
-			RealName: "test-screenshot",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-
-	req, err := http.NewRequest("GET", "/v2/find?q=test-screenshot", nil)
-	c.Assert(err, check.IsNil)
-	rsp, ok := searchStore(findCmd, req, nil).(*resp)
-	c.Assert(ok, check.Equals, true)
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-
-	c.Check(snaps[0]["name"], check.Equals, "test-screenshot")
-	c.Check(snaps[0]["media"], check.DeepEquals, []interface{}{
-		map[string]interface{}{
-			"type":   "screenshot",
-			"url":    "http://example.com/screenshot.png",
-			"width":  float64(800),
-			"height": float64(1280),
-		},
-		map[string]interface{}{
-			"type": "screenshot",
-			"url":  "http://example.com/screenshot2.png",
-		},
-	})
-}
-
 func (s *apiSuite) TestSnapsInfoOnlyStore(c *check.C) {
 	d := s.daemon(c)
 
@@ -1572,49 +1105,6 @@ func (s *apiSuite) TestSnapsInfoOnlyStore(c *check.C) {
 	c.Check(snaps[0]["prices"], check.IsNil)
 
 	c.Check(rsp.SuggestedCurrency, check.Equals, "EUR")
-}
-
-func (s *apiSuite) TestSnapsStoreConfinement(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = []*snap.Info{
-		{
-			// no explicit confinement in this one
-			SideInfo: snap.SideInfo{
-				RealName: "foo",
-			},
-		},
-		{
-			Confinement: snap.StrictConfinement,
-			SideInfo: snap.SideInfo{
-				RealName: "bar",
-			},
-		},
-		{
-			Confinement: snap.DevModeConfinement,
-			SideInfo: snap.SideInfo{
-				RealName: "baz",
-			},
-		},
-	}
-
-	req, err := http.NewRequest("GET", "/v2/find", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := searchStore(findCmd, req, nil).(*resp)
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 3)
-
-	for i, ss := range [][2]string{
-		{"foo", string(snap.StrictConfinement)},
-		{"bar", string(snap.StrictConfinement)},
-		{"baz", string(snap.DevModeConfinement)},
-	} {
-		name, mode := ss[0], ss[1]
-		c.Check(snaps[i]["name"], check.Equals, name, check.Commentf(name))
-		c.Check(snaps[i]["confinement"], check.Equals, mode, check.Commentf(name))
-	}
 }
 
 func (s *apiSuite) TestSnapsInfoStoreWithAuth(c *check.C) {
@@ -3579,17 +3069,6 @@ func (s *apiSuite) TestRevertSnapToRevisionJailMode(c *check.C) {
 
 func (s *apiSuite) TestRevertSnapToRevisionClassic(c *check.C) {
 	s.testRevertSnap(&snapInstruction{snapRevisionOptions: snapRevisionOptions{Revision: snap.R(1)}, Classic: true}, c)
-}
-
-func snapList(rawSnaps interface{}) []map[string]interface{} {
-	snaps := make([]map[string]interface{}, len(rawSnaps.([]*json.RawMessage)))
-	for i, raw := range rawSnaps.([]*json.RawMessage) {
-		err := json.Unmarshal([]byte(*raw), &snaps[i])
-		if err != nil {
-			panic(err)
-		}
-	}
-	return snaps
 }
 
 func (s *apiSuite) TestIsTrue(c *check.C) {
